--- src/js/declarations.td.ts
@@ -1,4 +1,0 @@
-declare module "~/data/option-values.json" {
-  const value: import("./types").OptionValues;
-  export default value;
-}
