+++ src/js/layout/fontAwesome.ts
