+++ tests/app/FilterState.test.ts
@@ -1,8 +1,11 @@
 import { expect, test } from "@playwright/test";
 
-import { FilterState, PlaceFilterManager } from "../../src/js/FilterState";
-import { POPULATION_MAX_INDEX } from "../../src/js/populationSlider";
-import { PlaceId, ProcessedCoreEntry, Date } from "../../src/js/types";
+import {
+  FilterState,
+  PlaceFilterManager,
+} from "../../src/js/state/FilterState";
+import { POPULATION_MAX_INDEX } from "../../src/js/filter-features/populationSlider";
+import { PlaceId, ProcessedCoreEntry, Date } from "../../src/js/model/types";
 
 test.describe("PlaceFilterManager.matchedPolicyRecords()", () => {
   const DEFAULT_STATE: FilterState = {
